feat: added delete, refac: finished git PR suggestions
