+++ src/components/Visit/visitinstance.tsx
@@ -1,172 +1,12 @@
 import React, { useState } from 'react'
 import ChevronDownIcon from '@heroicons/react/outline/ChevronDownIcon'
-import { updateUserData, User, Visit } from 'databaseintegration'
+import { Visit } from 'databaseintegration'
 
 import { EditButton } from './buttons'
+import EditableVisitInstance from './editvisitinstance'
+import ReadOnlyVisitInstance from './readvisitinstance'
 
 type VisitInstanceProps = Visit
-
-function NumberForm(value: string) {
-  if (isNaN(parseFloat(value))) {
-    return 0
-  }
-  return parseFloat(value)
-}
-
-const ReadOnlyVisitInstance = (props: User) => {
-  return (
-    <>
-      <div className='font-bold peer-checked:font-normal'>
-        <p className='font-bold text-primary'>{`${props.dateTime}`}</p>
-        <p className='text-sm'>{`${props.lastName}, ${props.firstName}`}</p>
-      </div>
-      <div className='max-h-0 justify-between overflow-hidden text-sm transition-all duration-300 peer-checked:max-h-40'>
-        <p>Pet/Pets: {props.petName}</p>
-        <p>Duration: {props.duration}</p>
-        <p>Walk Metres: {props.walkDist.toFixed(3)} km</p>
-        <p>Commute Metres: {props.commuteDist.toFixed(1)} km</p>
-        <p>Commute Method: {props.commuteMethod}</p>
-        <p>Notes: {props.notes}</p>
-      </div>
-    </>
-  )
-}
-
-const EditableVisitInstance = (props: Visit) => {
-  const [firstName, setFirstName] = useState(props.firstName)
-  const [lastName, setLastName] = useState(props.lastName)
-  const [dateTime, setDateTime] = useState(props.dateTime)
-  const [petName, setPetName] = useState(props.petName)
-  const [walkDist, setWalkDist] = useState(props.walkDist)
-  const [duration, setDuration] = useState(props.duration)
-  const [commuteDist, setCommuteDist] = useState(props.commuteDist)
-  const [commuteMethod, setCommuteMethod] = useState(props.commuteMethod)
-  const [notes, setNotes] = useState(props.notes)
-
-  return (
-    <form
-      onSubmit={(event) => {
-        const userData: User = {
-          firstName: firstName,
-          lastName: lastName,
-          petName: petName,
-          dateTime: dateTime,
-          duration: duration,
-          walkDist: walkDist,
-          commuteDist: commuteDist,
-          commuteMethod: commuteMethod,
-          notes: notes
-        }
-        updateUserData(props.id, userData)
-        event.preventDefault()
-      }}
-    >
-      <div className='font-bold peer-checked:font-normal'>
-        <input
-          placeholder='Date'
-          value={dateTime}
-          onChange={(event) => setDateTime(event.target.value)}
-          className='bg-gray font-normal text-primary'
-        />
-        <div>
-          <input
-            size={8}
-            className='bg-gray text-sm font-normal text-primary'
-            placeholder='Last Name'
-            value={lastName}
-            onChange={(event) => setLastName(event.target.value)}
-          />
-          <input
-            size={8}
-            className='bg-gray text-sm font-normal text-primary'
-            placeholder='First Name'
-            value={firstName}
-            onChange={(event) => setFirstName(event.target.value)}
-          />
-        </div>
-      </div>
-      <p className='text-sm'>
-        Pet/Pets:{' '}
-        <input
-          size={8}
-          className='bg-gray text-primary'
-          placeholder='Pet Name(s)'
-          value={petName}
-          onChange={(event) => setPetName(event.target.value)}
-        />
-      </p>
-      <p className='text-sm'>
-        Duration:{' '}
-        <input
-          size={8}
-          className='bg-gray text-primary'
-          placeholder='Duration'
-          value={duration}
-          onChange={(event) => setDuration(event.target.value)}
-        />
-      </p>
-      <p className='text-sm'>
-        Walk Metres:{' '}
-        <input
-          size={8}
-          className='bg-gray p-0 text-sm text-primary'
-          type='number'
-          step='0.001'
-          placeholder='Distance'
-          value={walkDist.toString()}
-          onChange={(event) => {
-            setWalkDist(NumberForm(event.target.value))
-          }}
-        />
-      </p>
-      <p className='text-sm'>
-        Commute Metres:{' '}
-        <input
-          size={8}
-          className='bg-gray text-primary'
-          type='number'
-          step='0.001'
-          placeholder='Distance'
-          value={commuteDist.toString()}
-          onChange={(event) => {
-            setCommuteDist(NumberForm(event.target.value))
-          }}
-        />
-      </p>
-      <p className='text-sm'>
-        Commute Method:{' '}
-        <input
-          size={8}
-          className='bg-gray text-primary'
-          placeholder='Method'
-          value={commuteMethod}
-          onChange={(event) => setCommuteMethod(event.target.value)}
-        />
-      </p>
-      <p className='text-sm'>
-        Notes:{' '}
-        <input // maybe use textarea tag instead?
-          size={8}
-          className='bg-gray text-primary'
-          placeholder='Notes'
-          value={notes}
-          onChange={(event) => setNotes(event.target.value)}
-        />
-      </p>
-      <button
-        type='submit'
-        className='text-bold mt-2 rounded-xl bg-primary p-2 text-white drop-shadow-default'
-        onClick={() =>
-          setTimeout(() => {
-            window.location.reload()
-          }, 50)
-        }
-      >
-        Submit
-      </button>
-    </form>
-  )
-}
 
 const VisitInstance = (props: VisitInstanceProps) => {
   const [isEditable, setIsEditable] = useState(false)
